// src/lib/exporters/ableton/types/returnTrack.ts
interface Name {
  EffectiveName: ValueElement;
  UserName: ValueElement;
  Annotation: ValueElement;
  MemorizedFirstClipName: ValueElement;
}

interface TrackDelay {
  Value: ValueElement;
  IsValueSampleBased: ValueElement;
}

interface AutomationEnvelopes {
  Envelopes: {};
}

interface TakeLanes {
  TakeLanes: {};
  AreTakeLanesFolded: ValueElement;
}

type EffectDevice = ALSChorus | ALSCompressor | ALSDelay | ALSDistortion | ALSFilter | ALSReverb;

interface DeviceChain {
  AutomationLanes: AutomationLanes;
  ClipEnvelopeChooserViewState: ClipEnvelopeChooserViewState;
  AudioInputRouting: Routing;
  MidiInputRouting: Routing;
  AudioOutputRouting: Routing;
  MidiOutputRouting: Routing;
  Mixer: Mixer;
  DeviceChain: {
    Devices: EffectDevice;
    SignalModulations: {};
  };
  FreezeSequencer: FreezeSequencer;
}

export interface ALSReturnTrackContent {
  '@Id': number;
  LomId: ValueElement;
  LomIdView: ValueElement;
  IsContentSelectedInDocument: ValueElement;
  PreferredContentViewMode: ValueElement;
  TrackDelay: TrackDelay;
  Name: Name;
  Color: ValueElement;
  AutomationEnvelopes: AutomationEnvelopes;
  TrackGroupId: ValueElement;
  TrackUnfolded: ValueElement;
  DevicesListWrapper: LomIdElement;
  ClipSlotsListWrapper: LomIdElement;
  ViewData: ValueElement;
  TakeLanes: TakeLanes;
  LinkedTrackGroupId: ValueElement;
  DeviceChain: DeviceChain;
}

export interface ALSReturnTrack {
  ReturnTrack: ALSReturnTrackContent;
}